+++ src/main/java/com/springboot/monew/users/controller/UserApiDocs.java
@@ -1,10 +1,10 @@
 package com.springboot.monew.users.controller;
 
 import com.springboot.monew.common.exception.ErrorResponse;
-import com.springboot.monew.users.dto.UserDto;
-import com.springboot.monew.users.dto.UserLoginRequest;
-import com.springboot.monew.users.dto.UserRegisterRequest;
-import com.springboot.monew.users.dto.UserUpdateRequest;
+import com.springboot.monew.users.dto.response.UserDto;
+import com.springboot.monew.users.dto.request.UserLoginRequest;
+import com.springboot.monew.users.dto.request.UserRegisterRequest;
+import com.springboot.monew.users.dto.request.UserUpdateRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
