+++ src/main/java/com/springboot/monew/users/mapper/UserMapper.java
@@ -1,6 +1,6 @@
 package com.springboot.monew.users.mapper;
 
-import com.springboot.monew.users.dto.UserDto;
+import com.springboot.monew.users.dto.response.UserDto;
 import com.springboot.monew.users.entity.User;
 import org.mapstruct.Mapper;
