+++ src/main/java/com/springboot/monew/users/repository/UserRepository.java
@@ -8,7 +8,6 @@ import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
-import org.springframework.stereotype.Repository;
 
 public interface UserRepository extends JpaRepository<User, UUID> {
